Split main CircleCI class into various files
